requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
openpyxl>=3.1.0
tqdm>=4.66.0
aiohttp>=3.9.0
asyncio>=3.4.3
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Optional: For JavaScript-rendered sites
# playwright>=1.40.0
# Install with: pip install playwright && playwright install

# Optional: For vision-based metadata extraction
# Pillow>=10.0.0
# pdf2image>=1.16.0
# PyMuPDF>=1.23.0
# Install with: pip install Pillow pdf2image PyMuPDF
# Note: pdf2image requires poppler-utils (Windows: download from https://github.com/oschwartz10612/poppler-windows/releases/)
